Extract composables from Testing Environment container
Split orchestration logic into three focused composables:
useToast, useTestingStatus, and useActions, leaving the
container as a thin composition surface.
