+++ openlibrary/components/TestingEnvironment.vue
@@ -1,25 +1,11 @@
 <script setup>
-import { shallowRef, computed, watch, onMounted, onBeforeUnmount } from 'vue';
+import { shallowRef, computed, watch, onBeforeUnmount } from 'vue';
 import TestingRow from './TestingEnvironment/TestingRow.vue';
 import DeploySection from './TestingEnvironment/DeploySection.vue';
-import {
-    DEFAULT_STRINGS,
-    applyDeployBadge,
-    decodeAndParseJSON,
-    effectiveActive,
-    faviconEnv,
-    getTestingStatus,
-    postAction,
-    sprintf
-} from './TestingEnvironment/utils.js';
-
-// The action endpoints answer {"ok": false, "error": "<code>"} for business
-// failures; map each code to the translated toast that explains it.
-const ACTION_ERRORS = {
-    add_failed: 'actionFailed',
-    deploy_failed: 'deployFailedTrigger',
-    deploy_unconfigured: 'deployUnconfigured'
-};
+import { DEFAULT_STRINGS, applyDeployBadge, decodeAndParseJSON, faviconEnv } from './TestingEnvironment/utils.js';
+import { useToast } from './TestingEnvironment/composables/useToast.js';
+import { useTestingStatus } from './TestingEnvironment/composables/useTestingStatus.js';
+import { useActions } from './TestingEnvironment/composables/useActions.js';
 
 defineOptions({ name: 'TestingEnvironment' });
 
@@ -40,27 +26,31 @@ const props = defineProps({
     }
 });
 
-// ── Reactive state ──────────────────────────────────────────────────
-const view = shallowRef('loading'); // 'loading' | 'error' | 'ready'
-const payload = shallowRef(null);
-const busy = shallowRef(false);
-const refreshing = shallowRef(false);
-const adding = shallowRef(false);
-const deploying = shallowRef(false);
-const addInput = shallowRef('');
+// ── i18n (runs at setup time — same timing as the old `created` hook) ─
 const strings = shallowRef({ ...DEFAULT_STRINGS });
-const toast = shallowRef('');
-// Wall-clock tick for the relative "X ago" deploy labels. Bumped by
-// the same poll that refreshes data, so the labels advance even when
-// the payload is unchanged (loadStatus skips identical JSON).
-const now = shallowRef(Date.now());
+try {
+    const parsed = decodeAndParseJSON(props.i18n);
+    if (parsed && typeof parsed === 'object' && !Array.isArray(parsed)) {
+        strings.value = { ...DEFAULT_STRINGS, ...parsed };
+    }
+} catch {
+    // A malformed translation payload falls back to English.
+}
 
-// ── Non-reactive instance state (plain let, not refs) ──────────────
-let timer = null;
-let toastTimer = null;
-let deployBadge = null;
+// ── Composables ─────────────────────────────────────────────────────
+const { toast, setToast } = useToast();
 
-// ── Computed ────────────────────────────────────────────────────────
+// busy is a shared re-entrancy guard between status fetching and actions.
+const busy = shallowRef(false);
+const { view, payload, now, loadStatus, retry } = useTestingStatus(busy);
+const { refreshing, adding, deploying, addInput, togglePr, updatePr, removePr, deploy, refresh, addPrs } = useActions({
+    busy,
+    loadStatus,
+    setToast,
+    strings
+});
+
+// ── Derived state ───────────────────────────────────────────────────
 const isMaintainer = computed(() => props.maintainer === 'true');
 
 const prs = computed(() => {
@@ -68,34 +58,9 @@ const prs = computed(() => {
     return ((payload.value && payload.value.prs) || []).filter((pr) => pr.merged !== true);
 });
 
-// ── Watchers ────────────────────────────────────────────────────────
-// The tab favicon follows the deploy: spinner-ring variant while a
-// build is presumed running, the normal one when it finishes.
-watch(
-    () => payload.value?.deploying,
-    (deploying) => syncDeployFavicon(deploying),
-    { immediate: true }
-);
+// ── Deploy favicon badge ────────────────────────────────────────────
+let deployBadge = null;
 
-// ── Methods ─────────────────────────────────────────────────────────
-function text(key, ...args) {
-    return sprintf(strings.value[key] || DEFAULT_STRINGS[key] || key, ...args);
-}
-
-function setToast(message) {
-    toast.value = message;
-    clearTimeout(toastTimer);
-    toastTimer = setTimeout(() => {
-        toast.value = '';
-    }, 6000);
-}
-
-// Mark the page favicon while a deploy runs: the real favicon is drawn
-// once with a static badge wedge and swapped into the rel="icon" links
-// (a static mark is one render at start and one swap at the end — no
-// animation loop for a throttled tab to stall). Only openlibrary
-// favicons are touched; the badge is removed when the deploy ends or
-// the panel unmounts.
 function syncDeployFavicon(isDeploying) {
     if (isDeploying) {
         if (deployBadge) return;
@@ -109,148 +74,13 @@ function syncDeployFavicon(isDeploying) {
     }
 }
 
-function onVisibilityChange() {
-    if (document.visibilityState === 'visible') {
-        now.value = Date.now();
-        silentRefresh();
-    }
-}
+watch(
+    () => payload.value?.deploying,
+    (deploying) => syncDeployFavicon(deploying),
+    { immediate: true }
+);
 
-// Re-fetch quietly: no loading view, no error takeover, no busy flag —
-// loadStatus(false, false, false). Skipped while an action is in flight.
-function silentRefresh() {
-    if (busy.value) return;
-    loadStatus(false, false, false);
-}
-
-async function loadStatus(showLoading = false, renderError = true, manageBusy = true) {
-    // busy is a re-entrancy guard and aria-busy signal only.
-    if (manageBusy) busy.value = true;
-    if (showLoading) view.value = 'loading';
-    try {
-        const newPayload = await getTestingStatus();
-        // Skip the assignment when nothing changed — a fresh object
-        // identity would repaint the panel (the flash on tab return).
-        if (!payload.value || JSON.stringify(newPayload) !== JSON.stringify(payload.value)) {
-            payload.value = newPayload;
-        }
-        view.value = 'ready';
-        return true;
-    } catch {
-        if (renderError) view.value = 'error';
-        return false;
-    } finally {
-        if (manageBusy) busy.value = false;
-    }
-}
-
-async function runAction(action, fields) {
-    if (busy.value) return false;
-    busy.value = true;
-    try {
-        const result = await postAction(action, fields);
-        await loadStatus(false, false, false);
-        // A business failure ({"ok": false, "error": "<code>"}) is a
-        // completed request, not a thrown fetch — say why instead of
-        // pretending the action landed.
-        if (result && result.ok === false) {
-            const key = ACTION_ERRORS[result.error] || 'actionFailed';
-            setToast(text(key));
-            return result;
-        }
-        return result;
-    } catch {
-        setToast(text('actionFailed'));
-        return false;
-    } finally {
-        busy.value = false;
-    }
-}
-
-function togglePr(pr) {
-    const action = effectiveActive(pr) ? '/status/disable' : '/status/enable';
-    runAction(action, { prs: [pr.pr] });
-}
-
-function updatePr(pr) {
-    runAction('/status/pull-latest', { prs: [pr.pr] });
-}
-
-function removePr(pr) {
-    runAction('/status/remove', { prs: [pr.pr] });
-}
-
-async function deploy() {
-    if (busy.value) return;
-    deploying.value = true;
-    try {
-        await runAction('/status/deploy', {});
-    } finally {
-        deploying.value = false;
-    }
-}
-
-async function refresh() {
-    if (busy.value) return;
-    refreshing.value = true;
-    try {
-        await runAction('/status/refresh', {});
-    } finally {
-        refreshing.value = false;
-    }
-}
-
-async function addPrs() {
-    if (adding.value || busy.value) return;
-    const value = addInput.value.trim();
-    if (!value) return;
-    adding.value = true;
-    try {
-        const result = await runAction('/status/add', { pr: value });
-        // A failed add keeps the input so it's obvious the PR didn't land.
-        if (result && result.ok) {
-            addInput.value = '';
-        }
-    } finally {
-        adding.value = false;
-    }
-}
-
-function retry() {
-    loadStatus(true);
-}
-
-// ── Lifecycle ───────────────────────────────────────────────────────
-// Parse the i18n translation payload (runs at setup time — same timing
-// as the old `created` hook).
-try {
-    const parsed = decodeAndParseJSON(props.i18n);
-    if (parsed && typeof parsed === 'object' && !Array.isArray(parsed)) {
-        strings.value = { ...DEFAULT_STRINGS, ...parsed };
-    }
-} catch {
-    // A malformed translation payload falls back to English.
-}
-
-onMounted(() => {
-    loadStatus();
-    // Single 1s interval: bumps `now` every tick (advances the label, no
-    // network), and refreshes data only on a tick at a :05 clock boundary.
-    timer = setInterval(() => {
-        now.value = Date.now();
-        if (Math.floor(now.value / 1000) % 5 === 0) {
-            silentRefresh();
-        }
-    }, 1000);
-    document.addEventListener('visibilitychange', onVisibilityChange);
-});
-
-onBeforeUnmount(() => {
-    clearTimeout(toastTimer);
-    clearInterval(timer);
-    document.removeEventListener('visibilitychange', onVisibilityChange);
-    syncDeployFavicon(false);
-});
+onBeforeUnmount(() => syncDeployFavicon(false));
 </script>
 
 <template>
